requirements.txt
langgraph
langchain-core
langchain-groq  # New dependency for Groq
python-dotenv
graphviz
rich